@@ -1,25 +1,80 @@
+import 'dart:convert';
 import 'package:flutter/foundation.dart';
 import 'package:flutter_tutorial_shop_app/providers/cart.dart';
+import 'package:http/http.dart' as http;
 
-class OrderItem{
+class OrderItem {
   final String id;
   final double amount;
   final List<CartItem> products;
   final DateTime dateTime;
 
-  OrderItem({@required this.id, @required this.amount, @required this.products, @required this.dateTime});
-
+  OrderItem(
+      {@required this.id,
+      @required this.amount,
+      @required this.products,
+      @required this.dateTime});
 }
 
-class Orders with ChangeNotifier{
+class Orders with ChangeNotifier {
   List<OrderItem> _ordres = [];
-
+  final String authToken;
+  Orders(this.authToken, this._ordres);
   List<OrderItem> get orders {
     return [..._ordres];
   }
 
-  void addOrder(List<CartItem> cartProducts, double total){
-    _ordres.insert(0, OrderItem(id: DateTime.now().toString(), amount: total, products: cartProducts, dateTime: DateTime.now()));
+  Future<void> fetchAndSetOrders() async {
+    final url =
+        'https://flutter-tutorial-shop.firebaseio.com/orders.json?auth=$authToken';
+    final response = await http.get(url);
+    final List<OrderItem> loadedOrders = [];
+    final extractedData = json.decode(response.body) as Map<String, dynamic>;
+    if (extractedData == null) {
+      return;
+    }
+    extractedData.forEach((orderId, orderData) {
+      loadedOrders.add(OrderItem(
+          id: orderId,
+          amount: orderData['amount'],
+          dateTime: DateTime.parse(orderData['dateTime']),
+          products: (orderData['products'] as List<dynamic>)
+              .map((item) => CartItem(
+                  id: item['id'],
+                  price: item['price'],
+                  quantity: item['quantity'],
+                  title: item['title']))
+              .toList()));
+    });
+    _ordres = loadedOrders.reversed.toList();
+    notifyListeners();
+  }
+
+  Future<void> addOrder(List<CartItem> cartProducts, double total) async {
+    final url =
+        'https://flutter-tutorial-shop.firebaseio.com/orders.json?auth=$authToken';
+    final timestamp = DateTime.now();
+    final response = await http.post(url,
+        body: json.encode({
+          'amount': total,
+          'dateTime': timestamp.toIso8601String(),
+          'products': cartProducts
+              .map((cp) => {
+                    'id': cp.id,
+                    'title': cp.title,
+                    'quantity': cp.quantity,
+                    'price': cp.price
+                  })
+              .toList()
+        }));
+
+    _ordres.insert(
+        0,
+        OrderItem(
+            id: json.decode(response.body)['name'],
+            amount: total,
+            products: cartProducts,
+            dateTime: DateTime.now()));
     //Anywhere that depends on the orders is now updated.
     notifyListeners();
   }
