@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_tutorial_shop_app/providers/auth.dart';
 import 'package:flutter_tutorial_shop_app/screens/cart_screen.dart';
 import 'package:flutter_tutorial_shop_app/screens/orders_screen.dart';
 import './screens/products_overview_screen.dart';
@@ -10,6 +11,7 @@ import 'package:provider/provider.dart';
 import './screens/orders_screen.dart';
 import './screens/user_product_screen.dart';
 import './screens/edit_product_screen.dart';
+import './screens/auth_screen.dart';
 
 void main() => runApp(MyApp());
 
@@ -17,26 +19,35 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MultiProvider(
-      providers: [
-        ChangeNotifierProvider.value(value: Products()),
-        ChangeNotifierProvider.value(value: Cart()),
-        ChangeNotifierProvider.value(value: Orders())
-      ],
-      child: MaterialApp(
-        title: 'MyShop',
-        theme: ThemeData(
-            primarySwatch: Colors.purple,
-            accentColor: Colors.deepOrange,
-            fontFamily: 'Lato'),
-        home: ProductsOverviewScreen(),
-        routes: {
-          EditProductScreen.routeName: (ctx) => EditProductScreen(),
-          ProductDetailScreen.routeName: (ctx) => ProductDetailScreen(),
-          CartScreen.routeName: (ctx) => CartScreen(),
-          OrdersScreen.routeName: (ctx) => OrdersScreen(),
-          UserProductScreen.routeName: (ctx) => UserProductScreen(),
-        },
-      ),
-    );
+        providers: [
+          ChangeNotifierProvider.value(value: Auth()),
+          ChangeNotifierProxyProvider<Auth, Products>(
+              builder: (ctx, auth, previousProducts) => Products(
+                  auth.token,
+                  auth.userId,
+                  previousProducts == null ? [] : previousProducts.items)),
+          ChangeNotifierProvider.value(value: Cart()),
+          ChangeNotifierProxyProvider<Auth, Orders>(
+              builder: (ctx, auth, previousOrders) => Orders(auth.token,
+                  previousOrders == null ? [] : previousOrders.orders))
+        ],
+        child: Consumer<Auth>(
+          builder: (ctx, auth, _) => MaterialApp(
+            title: 'MyShop',
+            theme: ThemeData(
+                primarySwatch: Colors.purple,
+                accentColor: Colors.deepOrange,
+                fontFamily: 'Lato'),
+            home: auth.isAuth ? ProductsOverviewScreen() : AuthScreen(),
+            routes: {
+              // ProductsOverviewScreen.rou
+              EditProductScreen.routeName: (ctx) => EditProductScreen(),
+              ProductDetailScreen.routeName: (ctx) => ProductDetailScreen(),
+              CartScreen.routeName: (ctx) => CartScreen(),
+              OrdersScreen.routeName: (ctx) => OrdersScreen(),
+              UserProductScreen.routeName: (ctx) => UserProductScreen(),
+            },
+          ),
+        ));
   }
 }
