@@ -11,6 +11,7 @@ class UserProductItem extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    final scaffold = Scaffold.of(context);
     return ListTile(
       title: Text(title),
       leading: CircleAvatar(
@@ -22,8 +23,12 @@ class UserProductItem extends StatelessWidget {
           IconButton(icon: Icon(Icons.edit), onPressed: () {
             Navigator.of(context).pushNamed(EditProductScreen.routeName, arguments: id);
           }, color: Theme.of(context).primaryColor,),
-          IconButton(icon: Icon(Icons.delete), onPressed: () {
+          IconButton(icon: Icon(Icons.delete), onPressed: ()  async {
+            try{
             Provider.of<Products>(context ,listen: false).deleteProduct(id);
+            } catch (error){
+              scaffold.showSnackBar(SnackBar(content: Center(child: Text('Deleting failed!')),));
+            }
           }, color: Theme.of(context).errorColor,),
         ],),
       ),
